# docs/swagger.yaml
basePath: /
definitions:
...
  properties.CommercialData:
    properties:
      ceiling_height:
        type: number
      internal_offices:
        type: integer
      land_use:
        type: string
      loading_docks:
        type: integer
      three_phase_power:
        type: boolean
    type: object
...
  properties.DeletePropertyInput:
    properties:
      confirm:
        example: true
        type: boolean
    type: object
...
  properties.GetPropertyData:
    properties:
      commercial:
        $ref: '#/definitions/properties.CommercialData'
      description:
        type: string
      is_featured:
        type: boolean
      location:
        $ref: '#/definitions/properties.LocationData'
      lot_area:
        type: number
      modality_id:
        type: integer
      property_type_id:
        type: integer
      property_uuid:
        type: string
      registered_by:
        type: string
      residential:
        $ref: '#/definitions/properties.ResidentialData'
      status_id:
        type: integer
      subtype:
        type: string
      title:
        type: string
    type: object
...
  properties.GetPropertyResult:
    properties:
      data:
        $ref: '#/definitions/properties.GetPropertyData'
    type: object
...
  properties.LocationData:
    properties:
      city_id:
        type: integer
      exterior_number:
        type: string
      interior_number:
        type: string
      is_public_address:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      neighborhood:
        type: string
      postal_code:
        type: string
      street:
        type: string
    type: object
...
  properties.ResidentialData:
    properties:
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      beds:
        type: integer
      built_area:
        type: number
      construction_year:
        type: integer
      floors:
        type: integer
      is_furnished:
        type: boolean
      orientation_id:
        type: integer
      parking_spots:
        type: integer
    type: object
  properties.UpdateCommercialInput:
    properties:
      ceiling_height:
        type: number
      internal_offices:
        type: integer
      land_use:
        type: string
      loading_docks:
        type: integer
      three_phase_power:
        type: boolean
    type: object
  properties.UpdateLocationInput:
    properties:
      city_id:
        type: integer
      exterior_number:
        type: string
      interior_number:
        type: string
      is_public_address:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      neighborhood:
        type: string
      postal_code:
        type: string
      street:
        type: string
    type: object
...
  properties.UpdatePropertyInput:
    properties:
      commercial:
        $ref: '#/definitions/properties.UpdateCommercialInput'
      description:
        type: string
      is_featured:
        type: boolean
      location:
        $ref: '#/definitions/properties.UpdateLocationInput'
      lot_area:
        type: number
      residential:
        $ref: '#/definitions/properties.UpdateResidentialInput'
      title:
        type: string
    type: object
...
  properties.UpdatePropertyResult:
    properties:
      message:
        type: string
    type: object
...
  properties.UpdateResidentialInput:
    properties:
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      beds:
        type: integer
      built_area:
        type: number
      construction_year:
        type: integer
      floors:
        type: integer
      is_furnished:
        type: boolean
      orientation_id:
        type: integer
      parking_spots:
        type: integer
    type: object
...
  shared.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API de Spazio Backend
  title: Spazio API
  version: "1.0"
paths:
...
  /api/v1/properties/{uuid}:
    delete:
      consumes:
      - application/json
      description: Soft deletes a property by UUID. Requires confirm=true. The authenticated
        admin is recorded as the author of the status change. The operation deletes
        linked photo objects from storage before applying database updates. Only properties
        with available status can be deleted.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Delete payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.DeletePropertyInput'
      produces:
      - application/json
      responses:
        "200":
          description: Property deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete property
      tags:
      - Properties
    get:
      description: Returns property base data, subtype, and location for the given
        UUID. Administrators can view all fields including registered_by. Agents can
        only access assigned properties and do not receive registered_by.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property data
          schema:
            $ref: '#/definitions/properties.GetPropertyResult'
        "400":
          description: Invalid path parameter
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "401":
          description: Missing or invalid authenticated session
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get property by UUID
      tags:
      - Properties
    patch:
      consumes:
      - application/json
      description: |-
        Updates property base data, subtype and location.
        Requires an authenticated admin session.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.UpdatePropertyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/properties.UpdatePropertyResult'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update property
      tags:
      - Properties
...
swagger: "2.0"